Cloud/api/requirements.txt
# Function dependencies, for example:
# package>=version

Flask==2.0.1
google-cloud-storage==1.38.0
numpy==1.19.5
keras==2.5.0rc0
tensorflow==2.5.0
pillow==7.2.0
click==8.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
opencv-python==4.4.0.44